Change deposit method from int to double.

+++ src/common/BankServerInterface.java
@@ -11,13 +11,13 @@ import domain.EditRecordFields;
 public interface BankServerInterface extends Remote
 {
 	//ManagerOperationsInterface
-	public Boolean createAccount(String firstName, String lastName, String address, String phone, String customerID,BranchID branch) throws RemoteException;
+	public Boolean createAccount(String firstName, String lastName, String address, String phone, String customerID, BranchID branch) throws RemoteException;
 	public Boolean editRecord(String customerID, EditRecordFields fieldName, String newValue) throws RemoteException;
 	public HashMap<String, String> getAccountCount() throws RemoteException;
 	
 	//ClientOperationsInterface
-	public void deposit(String customerID, int amt) throws RemoteException;
-	public void withdraw(String customerID, int amt) throws RemoteException;
+	public void deposit(String customerID, double amt) throws RemoteException;
+	public void withdraw(String customerID, double amt) throws RemoteException;
 	public double getBalance(String customerID) throws RemoteException;
 	
 	//MiscOperations
